devbuddy/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-sse==0.4.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
joblib==1.2.0
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.1
Pillow==9.5.0
prettyprint==0.1.5
pycparser==2.21
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
wrapt==1.15.0
